refactor: ProfilePage를 /profile 로 이동, 빌드 시 import 에러 해결
